+++ packages/typespec-go/test/local/nooptionalbody/zz_models.go
@@ -5,6 +5,6 @@
 package nooptionalbody
 
 type Widget struct {
-	// REQUIRED
+	// REQUIRED (transformed)
 	Weight *int32
 }