+++ packages/typespec-go/CHANGELOG.md
@@ -7,6 +7,13 @@
 * Fixed some cases where a client name could stutter.
 * Force the body paramter to be required for `PATCH` and `PUT` operations.
 
+### Features Added
+
+* When Go tools are found on the path, the following steps happen after successfully generating code.
+  * Execute `gofmt -w .`.
+  * If `/internal/generate/transforms.go` exists in the output directory, `go run` it from the output directory.
+  * If Go tools are not found, the above steps are skipped and a warning is displayed.
+
 ## 0.5.1 (2025-06-26)
 
 ### Other Changes